Extract common invokable prim code
