+++ src/som/primitives/bc/method_primitives.py
@@ -1,16 +1,4 @@
-from som.primitives.primitives import Primitives
-from som.vmobjects.primitive   import UnaryPrimitive
+from som.primitives.invokable_primitives import InvokablePrimitivesBase as _Base
 
 
-def _holder(rcvr):
-    return rcvr.get_holder()
-
-
-def _signature(rcvr):
-    return rcvr.get_signature()
-
-
-class MethodPrimitives(Primitives):
-    def install_primitives(self):
-        self._install_instance_primitive(UnaryPrimitive("holder", self._universe, _holder))
-        self._install_instance_primitive(UnaryPrimitive("signature", self._universe, _signature))
+MethodPrimitives = _Base
